[feat] (bitcoinish) Move getTransactionInfo to utils and return outputUtxos
