+++ packages/bitcoin-payments/src/bitcoinish/BitcoinishPayments.ts
@@ -855,83 +855,6 @@ export abstract class BitcoinishPayments<Config extends BaseConfig> extends Bitc
     }
   }
 
-  async getTransactionInfo(txId: string): Promise<BitcoinishTransactionInfo> {
-    const tx = await this._retryDced(() => this.getApi().getTx(txId))
-    const txSpecific = await this._retryDced(() => this.getApi().getTxSpecific(txId))
-
-    // Our "weight" for fee purposes is vbytes, but that isn't a thing on all networks (BCH, DOGE)
-    const weight = txSpecific.vsize || txSpecific.size
-
-    const fee = this.toMainDenominationString(tx.fees)
-
-    const currentBlockNumber = (await this.getApi().getStatus()).blockbook.bestHeight
-
-    const confirmationId = tx.blockHash || null
-    const confirmationNumber = tx.blockHeight ? String(tx.blockHeight) : undefined
-    const confirmationTimestamp = tx.blockTime ? new Date(tx.blockTime * 1000) : null
-    if (tx.confirmations > 0x7FFFFFFF) {
-      // If confirmations exceeds the max value of a signed 32 bit integer, assume we have bad data
-      // Blockbook sometimes returns a confirmations count equal to `0xFFFFFFFF`
-      // Bitcoin won't have that many confirmations for 40,000 years
-      throw new Error(`Blockbook returned confirmations count for tx ${txId} that's way too big to be real (${tx.confirmations})`)
-    }
-    const isConfirmed = Boolean(tx.confirmations && tx.confirmations > 0)
-    const status = isConfirmed ? TransactionStatus.Confirmed : TransactionStatus.Pending
-    const inputUtxos = tx.vin.map(({ txid, vout, value }): UtxoInfo => ({
-      txid: txid || '',
-      vout: vout || 0,
-      value: this.toMainDenominationString(value || 0),
-    }))
-    const fromAddress = this.standardizeAddress(get(tx, 'vin.0.addresses.0', ''))
-    if (!fromAddress) {
-      throw new Error(`Unable to determine fromAddress of ${this.coinSymbol} tx ${txId}`)
-    }
-
-    const externalOutputs = tx.vout
-      .map(({ addresses, value }): TransactionOutput => (
-        {
-          address: this.standardizeAddress(addresses?.[0] ?? '') || '',
-          value: this.toMainDenominationString(value || 0),
-        }
-      ))
-      .filter(({ address }) => address !== fromAddress)
-    const amount = externalOutputs.reduce((total, { value }) => total.plus(value), new BigNumber(0))
-
-    let toAddress
-    if (externalOutputs.length > 1) {
-      toAddress = 'multioutput'
-    } else if (externalOutputs.length === 1) {
-      toAddress = externalOutputs[0].address
-    } else {
-      throw new Error(`Unable to determine toAddress of ${this.coinSymbol} tx ${txId}`)
-    }
-
-    return {
-      status,
-      id: tx.txid,
-      fromIndex: null,
-      fromAddress,
-      fromExtraId: null,
-      toIndex: null,
-      toAddress,
-      toExtraId: null,
-      amount: amount.toFixed(),
-      fee,
-      sequenceNumber: null,
-      confirmationId,
-      confirmationNumber,
-      currentBlockNumber,
-      confirmationTimestamp,
-      isExecuted: isConfirmed,
-      isConfirmed,
-      confirmations: tx.confirmations,
-      data: tx,
-      inputUtxos,
-      externalOutputs,
-      weight,
-    }
-  }
-
   private prepareUtxos(utxos: UtxoInfo[]): Array<UtxoInfo & { satoshis: number }> {
     return utxos.map((utxo) => {
       return {
